+++ lib/src/languages/java/java8/java8_grammar.dart
@@ -140,8 +140,14 @@ class Java8GrammarDefinition extends Java8GrammarLexer {
         return all[0] as ASTExpressionOperation;
       });
 
-  Parser<ASTExpressionOperator> expressionOperator() =>
-      (char('+') | char('-') | char('*') | char('/')).trim().map((v) {
+  Parser<ASTExpressionOperator> expressionOperator() => (char('+') |
+              char('-') |
+              char('*') |
+              char('/') |
+              string('==') |
+              string('!='))
+          .trim()
+          .map((v) {
         return getASTExpressionOperator(v);
       });